# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /security-info:
    get:
      tags:
        - security-info-controller
      description: List of all users security info in DB
      operationId: getAll_1
      responses:
        "200":
          description: "For admin: List of all users in DB (JSON array)."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockUser'
        "403":
          description: "For user: User cannot see all security info."
      security:
        - Bearer Authentication: []
    put:
      tags:
        - security-info-controller
      description: Update security info for user.
      operationId: update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityInfo'
        required: true
      responses:
        "400":
          description: Email or username is already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "403":
          description: "For user: If user wants to update another users security info."
        "404":
          description: "For administrator: No such user."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "200":
          description: Users security info is successfully changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityInfo'
      security:
        - Bearer Authentication: []
    post:
      tags:
        - security-info-controller
      description: Creating a user
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityInfo'
        required: true
      responses:
        "400":
          description: Not enough data for registration OR Email or username is already
            used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "403":
          description: "For user: If user wants to create."
        "201":
          description: Security info is successfully created in DB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityInfo'
        "404":
          description: No User with this ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
      security:
        - Bearer Authentication: []
...
  /security-info/{userId}:
    get:
      tags:
        - security-info-controller
      description: Get security info for particular user.
      operationId: getById_1
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "404":
          description: "For administrator: No such user in DB."
          content:
            application.json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "403":
          description: "For user: If user wants to access another users security info."
        "200":
          description: User is found in DB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityInfo'
      security:
        - Bearer Authentication: []
    delete:
      tags:
        - security-info-controller
      description: Deleting a Security info for User
      operationId: delete
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "204":
          description: Security info is successfully deleted
        "404":
          description: "For admin: No such user."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        "403":
          description: "For user: If user wants to delete another users security info."
      security:
        - Bearer Authentication: []
components:
  schemas:
    StockUser:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        firstName:
          type: string
        secondName:
          type: string
        birthday:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: boolean
    ErrorModel:
      type: object
      properties:
        error:
          type: string
    GrantedAuthority:
      type: object
      properties:
        authority:
          type: string
    SecurityInfo:
      type: object
      properties:
        userId:
          type: integer
          format: int32
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        enabled:
          type: boolean
        authorities:
          type: array
          items:
            $ref: '#/components/schemas/GrantedAuthority'
        credentialsNonExpired:
          type: boolean
        accountNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
...
  securitySchemes:
    Bearer Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT